+++ edx_proctoring/tests/test_mfe_views.py
@@ -10,13 +10,14 @@ from django.conf import settings
 from django.test.utils import override_settings
 from django.urls import reverse
 
+from edx_proctoring.api import get_review_policy_by_exam_id
 from edx_proctoring.exceptions import BackendProviderNotConfigured, ProctoredExamNotFoundException
 from edx_proctoring.statuses import ProctoredExamStudentAttemptStatus
 
 from .utils import ProctoredExamTestCase
 
 
-@override_settings(LEARNING_MICROFRONTEND_URL='https//learningmfe')
+@override_settings(LEARNING_MICROFRONTEND_URL='https//learningmfe', ACCOUNT_MICROFRONTEND_URL='https//localhost')
 @ddt.ddt
 class ProctoredExamAttemptsMFEViewTests(ProctoredExamTestCase):
     """
@@ -29,6 +30,8 @@ class ProctoredExamAttemptsMFEViewTests(ProctoredExamTestCase):
         super().setUp()
         self.timed_exam_id = self._create_timed_exam()
         self.proctored_exam_id = self._create_proctored_exam()
+        self.practice_exam_id = self._create_practice_exam()
+        self.onboarding_exam_id = self._create_onboarding_exam()
         self.url = reverse(
             'edx_proctoring:proctored_exam.exam_attempts',
             kwargs={
@@ -40,13 +43,20 @@ class ProctoredExamAttemptsMFEViewTests(ProctoredExamTestCase):
             settings.LEARNING_MICROFRONTEND_URL, self.course_id, self.content_id
         )
 
-    def assertHasExamData(self, response_data, has_attempt, content_id=None):
+    def assertHasExamData(self, response_data, has_attempt,
+                          has_verification_url=False, has_download_url=False, content_id=None):
         """ Ensure expected exam data is present. """
         exam_data = response_data['exam']
         assert 'exam' in response_data
         assert 'attempt' in exam_data
         if has_attempt:
             assert exam_data['attempt']
+            if has_verification_url:
+                assert exam_data['attempt']['verification_url']
+            if has_download_url:
+                assert 'software_download_url' in exam_data['attempt']
+            else:
+                assert 'software_download_url' not in exam_data['attempt']
         else:
             assert not exam_data['attempt']
         self.assertEqual(exam_data['course_id'], self.course_id)
@@ -63,9 +73,13 @@ class ProctoredExamAttemptsMFEViewTests(ProctoredExamTestCase):
         self.assertEqual(response.status_code, 200)
         response_data = json.loads(response.content.decode('utf-8'))
         exam_data = response_data['exam']
+
+        # if exam started the prerequisites are not checked
+        assert 'prerequisite_status' not in exam_data
         assert 'active_attempt' in response_data and response_data['active_attempt']
         self.assertHasExamData(response_data, has_attempt=True)
         self.assertEqual(exam_data['attempt']['exam_url_path'], self.expected_exam_url)
+        self.assertEqual(exam_data['type'], 'proctored')
 
     def test_get_started_timed_exam_attempts_data(self):
         """
@@ -88,19 +102,23 @@ class ProctoredExamAttemptsMFEViewTests(ProctoredExamTestCase):
         expected_exam_url = '{}/course/{}/{}'.format(
             settings.LEARNING_MICROFRONTEND_URL, self.course_id, self.content_id_timed
         )
+        assert 'prerequisite_status' not in exam_data
         assert 'active_attempt' in response_data and response_data['active_attempt']
         self.assertHasExamData(response_data, has_attempt=True, content_id=self.content_id_timed)
         self.assertEqual(exam_data['attempt']['exam_url_path'], expected_exam_url)
 
     def test_no_attempts_data_before_exam_starts(self):
         """
-        Test we get exam data before exam is started. Ensure no attempts data returned.
+        Test we get exam data before exam is started. Ensure no attempts data returned and prerequisites are checked.
         """
         response = self.client.get(self.url)
         self.assertEqual(response.status_code, 200)
         response_data = json.loads(response.content.decode('utf-8'))
+        exam_data = response_data['exam']
         assert 'active_attempt' in response_data and not response_data['active_attempt']
+        assert 'prerequisite_status' in exam_data
         self.assertHasExamData(response_data, has_attempt=False)
+        self.assertEqual(exam_data['type'], 'proctored')
 
     def test_get_exam_attempts_data_after_exam_is_submitted(self):
         """
@@ -113,9 +131,11 @@ class ProctoredExamAttemptsMFEViewTests(ProctoredExamTestCase):
         response_data = json.loads(response.content.decode('utf-8'))
         exam_data = response_data['exam']
         assert 'active_attempt' in response_data and not response_data['active_attempt']
+        assert 'prerequisite_status' not in exam_data
         self.assertHasExamData(response_data, has_attempt=True)
         self.assertEqual(exam_data['attempt']['exam_url_path'], self.expected_exam_url)
         self.assertEqual(exam_data['attempt']['attempt_status'], ProctoredExamStudentAttemptStatus.submitted)
+        self.assertEqual(exam_data['type'], 'proctored')
 
     def test_no_exam_data_returned_for_non_exam_sequence(self):
         """
@@ -135,6 +155,66 @@ class ProctoredExamAttemptsMFEViewTests(ProctoredExamTestCase):
         assert 'active_attempt' in response_data
         assert not response_data['active_attempt']
         assert not exam_data
+
+    @ddt.data(
+        ('content_id', True),
+        ('content_id_timed', False),
+        ('content_id_onboarding', False),
+        ('content_id_practice', False),
+    )
+    @ddt.unpack
+    def test_prerequisites_are_not_checked_if_exam_is_not_proctored(self, content_id, should_check_prerequisites):
+        """
+        Tests that prerequisites are not checked for non proctored exams.
+        """
+        content_id = getattr(self, content_id)
+        url = reverse(
+            'edx_proctoring:proctored_exam.exam_attempts',
+            kwargs={
+                'course_id': self.course_id,
+                'content_id': content_id
+            }
+        ) + '?is_learning_mfe=true'
+
+        response = self.client.get(url)
+        self.assertEqual(response.status_code, 200)
+        response_data = json.loads(response.content.decode('utf-8'))
+        exam_data = response_data['exam']
+        self.assertEqual('prerequisite_status' in exam_data, should_check_prerequisites)
+
+    @ddt.data(
+        ProctoredExamStudentAttemptStatus.created,
+        ProctoredExamStudentAttemptStatus.download_software_clicked,
+        ProctoredExamStudentAttemptStatus.ready_to_start,
+        ProctoredExamStudentAttemptStatus.started,
+        ProctoredExamStudentAttemptStatus.ready_to_submit,
+        ProctoredExamStudentAttemptStatus.declined,
+        ProctoredExamStudentAttemptStatus.submitted,
+        ProctoredExamStudentAttemptStatus.rejected,
+        ProctoredExamStudentAttemptStatus.expired
+    )
+    def test_exam_data_contains_necessary_data_based_on_the_attempt_status(self, status):
+        """
+        Tests the GET exam attempts data contains software download url ONLY when attempt
+        is in created or download_software_clicked status and contains verification
+        url ONLY when attempt is in created status
+        """
+        self._create_exam_attempt(self.proctored_exam_id, status=status)
+
+        response = self.client.get(self.url)
+        self.assertEqual(response.status_code, 200)
+        has_download_url = status in (
+            ProctoredExamStudentAttemptStatus.created,
+            ProctoredExamStudentAttemptStatus.download_software_clicked
+        )
+        has_verification_url = status == ProctoredExamStudentAttemptStatus.created
+        response_data = json.loads(response.content.decode('utf-8'))
+        self.assertHasExamData(
+            response_data,
+            has_attempt=True,
+            has_download_url=has_download_url,
+            has_verification_url=has_verification_url,
+        )
 
 
 class ProctoredSettingsViewTests(ProctoredExamTestCase):
@@ -202,3 +282,45 @@ class ProctoredSettingsViewTests(ProctoredExamTestCase):
             response = self.client.get(url)
         self.assertEqual(response.status_code, 500)
         self.assertRaises(BackendProviderNotConfigured)
+
+
+class ProctoredExamReviewPolicyView(ProctoredExamTestCase):
+    """
+    Tests for the ProctoredExamReviewPolicyView.
+    """
+
+    def setUp(self):
+        """
+        Initialize.
+        """
+        super().setUp()
+        self.proctored_exam_id = self._create_proctored_exam()
+        self.review_policy_id = self._create_review_policy(self.proctored_exam_id)
+        self.url = reverse(
+            'edx_proctoring:proctored_exam.review_policy',
+            kwargs={
+                'exam_id': self.proctored_exam_id,
+            }
+        )
+
+    def test_get_exam_review_policy_for_proctored_exam(self):
+        """
+        Tests the GET exam review policy endpoint for proctored exam with existing policy.
+        """
+        response = self.client.get(self.url)
+        self.assertEqual(response.status_code, 200)
+        response_data = json.loads(response.content.decode('utf-8'))
+        expected_review_policy = get_review_policy_by_exam_id(self.proctored_exam_id)
+        assert 'review_policy' in response_data
+        self.assertEqual(response_data['review_policy'], expected_review_policy['review_policy'])
+
+    def test_get_exam_review_policy_for_proctored_exam_with_no_existing_review(self):
+        """
+        Tests the GET exam review policy endpoint for proctored exam which has no review policy configured.
+        """
+        with patch('edx_proctoring.models.ProctoredExamReviewPolicy.get_review_policy_for_exam', return_value=None):
+            response = self.client.get(self.url)
+        self.assertEqual(response.status_code, 200)
+        response_data = json.loads(response.content.decode('utf-8'))
+        assert 'review_policy' in response_data
+        self.assertIsNone(response_data['review_policy'])