+++ edx_proctoring/utils.py
@@ -302,11 +302,22 @@ def get_visibility_check_date(course_schedule, usage_key):
     return visibility_check_date
 
 
-def get_exam_type(provider, attempt):
-    """ Helper that returns exam type string by backend provider and exam attempt params. """
-    return (
-        _('a timed exam') if not attempt['taking_as_proctored'] else
-        (_('a proctored exam') if not attempt['is_sample_attempt'] else
-         (_('an onboarding exam') if (provider and provider.supports_onboarding) else
-          _('a practice exam')))
-    )
+def get_exam_type(exam, provider):
+    """ Helper that returns exam type and humanized name by backend provider and exam params. """
+    is_practice_exam = exam['is_proctored'] and exam['is_practice_exam']
+    is_timed_exam = not exam['is_proctored'] and not exam['is_practice_exam']
+
+    exam_type, humanized_type = 'proctored', _('a proctored exam')
+
+    if is_timed_exam:
+        exam_type, humanized_type = 'timed', _('a timed exam')
+    elif is_practice_exam:
+        if provider and provider.supports_onboarding:
+            exam_type, humanized_type = 'onboarding', _('an onboarding exam')
+        else:
+            exam_type, humanized_type = 'practice', _('a practice exam')
+
+    return {
+        'type': exam_type,
+        'humanized_type': humanized_type,
+    }